Add a Twilio client implementation to internal/atlas-alerter/notify and remove twilio-go
- Sync `vendor/` for unused dependencies.
- All test variations pass.

+++ pkg/decode/chirpstack/device/device_join_test.go
@@ -93,7 +93,7 @@ func TestDeviceJoin(t *testing.T) {
 				{Attr: "devaddr", Value: devAddr},
 				{Attr: "gateway_id", Value: gatewayID},
 				{Attr: "time", Value: strconv.FormatInt(now.Unix(), 10)},
-				{Attr: "lora_rssi", Value: -74},
+				{Attr: "lora_rssi", Value: int32(-74)},
 				{Attr: "snr", Value: 7.8},
 				{Attr: "channel", Value: int32(0)},
 				{Attr: "frequency", Value: int32(902700000)},
